requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
protobuf==5.27.2
requests==2.32.3
urllib3==2.2.2
